service work

+++ src/main/java/com/sexyalbum/jdbc/UserDao.java
@@ -5,7 +5,7 @@ import com.sexyalbum.model.User;
 import java.util.List;
 
 public interface UserDao {
-    int add(User user);
+    Long add(User user);
     int update(User user);
     int delete(Long userid);
     User find(Long userid);
